Create pong_game.py
Most of us have heard about the famous pong game. Many of us love playing it. Today lets learn how to code this classic game using the python programming language!

Before starting with the coding part we first need to install the turtle module. The turtle module is a Python library that enables users to create pictures and shapes by providing them with a virtual canvas.